# Firework colors and Color API revamp