--- a/src/main/java/ch/njol/skript/command/CommandHelp.java
+++ b/src/main/java/ch/njol/skript/command/CommandHelp.java
@@ -57,14 +57,14 @@ public class CommandHelp {
 	
 	public CommandHelp(final String command, final Color argsColor, final String langNode) {
 		this.command = command;
-		this.argsColor = "" + argsColor.getChat();
+		this.argsColor = "" + argsColor.getFormattedChat();
 		this.langNode = langNode;
 		description = new Message(langNode + "." + DEFAULTENTRY);
 	}
 	
 	public CommandHelp(final String command, final Color argsColor) {
 		this.command = command;
-		this.argsColor = "" + argsColor.getChat();
+		this.argsColor = "" + argsColor.getFormattedChat();
 	}
 	
 	public CommandHelp add(final String argument) {