--- a/src/main/java/ch/njol/skript/entity/SheepData.java
+++ b/src/main/java/ch/njol/skript/entity/SheepData.java
@@ -20,6 +20,7 @@
 package ch.njol.skript.entity;
 
 import java.util.Arrays;
+import java.util.Optional;
 
 import org.bukkit.entity.Sheep;
 import org.eclipse.jdt.annotation.Nullable;
@@ -31,6 +32,7 @@ import ch.njol.skript.localization.Adjective;
 import ch.njol.skript.localization.Language;
 import ch.njol.skript.localization.Noun;
 import ch.njol.skript.util.Color;
+import ch.njol.skript.util.SkriptColor;
 import ch.njol.util.Checker;
 import ch.njol.util.coll.CollectionUtils;
 
@@ -43,7 +45,7 @@ public class SheepData extends EntityData<Sheep> {
 	}
 	
 	@Nullable
-	private Color[] colors = null;
+	private Color[] colors;
 	private int sheared = 0;
 	
 	@SuppressWarnings("unchecked")
@@ -57,9 +59,10 @@ public class SheepData extends EntityData<Sheep> {
 	
 	@SuppressWarnings("null")
 	@Override
-	protected boolean init(final @Nullable Class<? extends Sheep> c, final @Nullable Sheep e) {
+	protected boolean init(@Nullable Class<? extends Sheep> c, @Nullable Sheep e) {
 		sheared = e == null ? 0 : e.isSheared() ? 1 : -1;
-		colors = e == null ? null : new Color[] {Color.byWoolColor(e.getColor())};
+		Optional<SkriptColor> color = SkriptColor.fromDyeColor(e.getColor());
+		colors = (e == null || !color.isPresent()) ? null : new Color[] {color.get()};
 		return true;
 	}
 	
@@ -68,7 +71,7 @@ public class SheepData extends EntityData<Sheep> {
 		if (colors != null) {
 			final Color c = CollectionUtils.getRandom(colors);
 			assert c != null;
-			entity.setColor(c.getWoolColor());
+			entity.setColor(c.asDyeColor());
 		}
 	}
 	
@@ -78,7 +81,7 @@ public class SheepData extends EntityData<Sheep> {
 				&& (colors == null || SimpleExpression.check(colors, new Checker<Color>() {
 					@Override
 					public boolean check(final @Nullable Color c) {
-						return c != null && entity.getColor() == c.getWoolColor();
+						return c != null && entity.getColor() == c.asDyeColor();
 					}
 				}, false, false));
 	}
@@ -157,7 +160,7 @@ public class SheepData extends EntityData<Sheep> {
 					final String c = cs[i];
 					assert c != null;
 					assert colors != null;
-					colors[i] = Color.valueOf(c);
+					colors[i] = SkriptColor.valueOf(c);
 				} catch (final IllegalArgumentException e) {
 					return false;
 				}