… src/main/java/ch/njol/skript/util/Color.java
… src/main/java/ch/njol/skript/util/ColorRGB.java
… src/main/java/ch/njol/skript/util/SkriptColor.java
… src/main/java/ch/njol/skript/util/Utils.java
… src/main/java/ch/njol/skript/util/chat/ChatMessages.java
… src/main/resources/lang/english.lang
… src/test/java/ch/njol/skript/registrations/ClassesTest.java