Related issues: https://github.com/SkriptLang/Skript/issues/206 https://github.com/SkriptLang/Skript/issues/1291

Description:
- Adds exactly the same syntax as SkQuery. The coding being modified from my fork.
- This addition has been thoroughly tested.
- I understand that this could be implemented in aliases, but this is much cleaner, infinite options, friendly syntax, faster and overall a better implementation.
- Works with my SkQuery fork enabled, my fork has a classinfo checker so it doesn't error when the same classinfo is registered before it (aka Skript).
- I could not add the "pop firework" effect which essentially explodes the firework at the direct location specified, this is due to the usage of reflection, SkQuery does this.

Test script used:
```
command /firework:
	trigger:
		set {_location} to player's location
		set {_effect} to flickering ball large firework effect coloured light green fading to green, purple and white
		launch {_effect} at {_location} timed 1
		wait 7 seconds
		set {_effect} to ball firework effect coloured red, purple and light green
		launch {_effect} at {_location}
		wait 7 seconds
		set {_effect} to trailing creeper firework coloured red fading to white
		launch {_effect} at {_location} timed 2
		wait 7 seconds
		set {_effect} to flickering trailing star firework colored yellow
		launch {_effect} at {_location} timed 3
		wait 7 seconds
		set {_effect} to trailing flickering ball large coloured blue
		launch {_effect} at {_location}
		wait 7 seconds
		set {_effect} to creeper firework coloured pink and black
		launch {_effect} at {_location} timed 0
		wait 7 seconds
```

Videos:
(Old video from when I fixed the SkQuery fireworks): https://www.youtube.com/watch?v=EKYz9KcfCA8
(Test script video): https://youtu.be/CfdczdnmXXU