@TheLimeGlass 2018-11-03
For some reason I could not merge my branch `firework-colors` into `firework` I think it was from the limiting issue this week from @bensku 

I have made a new thread for this, which should be better and cleaner now.

Old issue https://github.com/SkriptLang/Skript/pull/1629