@TheLimeGlass 2018-11-03
This pull also takes into consideration https://github.com/SkriptLang/Skript/pull/1639