@TheLimeGlass 2019-01-26
I have tested this to an extreme extent when I originally made this by the way.